feat: equalizer handler

+++ src/equalizer-v3/v3-core/contracts/interfaces/IEqualizerV3Factory.sol
@@ -16,11 +16,7 @@ interface IEqualizerV3Factory {
     /// @param tickSpacing The minimum number of ticks between initialized ticks
     /// @param pool The address of the created pool
     event PoolCreated(
-        address indexed token0,
-        address indexed token1,
-        uint24 indexed fee,
-        int24 tickSpacing,
-        address pool
+        address indexed token0, address indexed token1, uint24 indexed fee, int24 tickSpacing, address pool
     );
 
     /// @notice Emitted when a new fee amount is enabled for pool creation via the factory
@@ -45,21 +41,17 @@ interface IEqualizerV3Factory {
     /// @param tokenB The contract address of the other token
     /// @param fee The fee collected upon every swap in the pool, denominated in hundredths of a bip
     /// @return pool The pool address
-    function getPool(
-        address tokenA,
-        address tokenB,
-        uint24 fee
-    ) external view returns (address pool);
+    function getPool(address tokenA, address tokenB, uint24 fee) external view returns (address pool);
 
     /// @notice Returns the pool address at input index, or address 0 if it does not exist
     /// @param index The chronological serial number of a Thick Pool
     /// @return pool The pool address
-    function allPairs(uint index) external view returns(address pool);
-/*
+    function allPairs(uint256 index) external view returns (address pool);
+    /*
     /// @notice Returns the total number of Thick Liquidity pools in existence
     /// @return pools The pool address
     function allPairsLength() external view returns(uint pools);
-*/
+    */
     /// @notice Gets the initialization code hash of a Thick Pool
     /// @dev Should be keccak256 of a Thick Pool's creationCode
     /// @return hash The default Protocol Fees denominator
@@ -73,11 +65,7 @@ interface IEqualizerV3Factory {
     /// from the fee. The call will revert if the pool already exists, the fee is invalid, or the token arguments
     /// are invalid.
     /// @return pool The address of the newly created pool
-    function createPool(
-        address tokenA,
-        address tokenB,
-        uint24 fee
-    ) external returns (address pool);
+    function createPool(address tokenA, address tokenB, uint24 fee) external returns (address pool);
 
     /// @notice Updates the owner of the factory
     /// @dev Must be called by the current owner
@@ -99,5 +87,4 @@ interface IEqualizerV3Factory {
     /// @dev Should be 0-255.
     /// @param fee The default Protocol Fees denominator
     function setFeeProtocol(uint8 fee) external;
-
 }
